refactor: LottoService 분리
LottoService를 LottoChecker와 LottoGenerator로 분리
LottoService는 단순히 분리된 클래스를 호출하고 값을 전달해주는 역할만 한다.
